# rest/openapi.yaml
...
openapi: 3.1.0
info:
    title: Taqueria
    version: 1.0.0
servers:
    - url: http://localhost:8080/v1
paths:
...
    /taquerias:
        get:
            tags:
                - Taqueria
            summary: List taquerias
            operationId: listTaqueria
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of taquerias
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TaqueriaList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - Taqueria
            summary: Create Taqueria
            operationId: createTaqueria
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Taqueria'
                required: true
            responses:
                "201":
                    description: Created Taqueria
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Taqueria'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
    /visitas:
        get:
            tags:
                - Visita
            summary: List visitas
            operationId: listVisita
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of visitas
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/VisitaList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - Visita
            summary: Create Visita
            operationId: createVisita
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Visita'
                required: true
            responses:
                "201":
                    description: Created Visita
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Visita'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
components:
    schemas:
        Problem:
            properties:
                detail:
                    type: string
                instance:
                    type: string
                status:
                    type: integer
                title:
                    type: string
                type:
                    type: string
            type: object
...
        Taqueria:
            properties:
                colonia:
                    type: string
                created_at:
                    format: date-time
                    type: string
                especialidad:
                    $ref: '#/components/schemas/TipoTaco'
                id:
                    format: uuid
                    type: string
                name:
                    type: string
                updated_at:
                    format: date-time
                    type: string
            required:
                - id
                - name
                - colonia
                - especialidad
                - created_at
                - updated_at
            type: object
        TaqueriaList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/Taqueria'
                    type: array
                next_page_token:
                    type: string
            type: object
        TipoTaco:
            enum:
                - invalid
                - pastor
                - suadero
                - campechano
                - bistec
                - lengua
                - tripa
                - chorizo
                - canasta
            type: string
        Veredicto:
            enum:
                - invalid
                - imperdible
                - bueno
                - pasable
                - nunca_mas
            type: string
        Visita:
            properties:
                created_at:
                    format: date-time
                    type: string
                fecha:
                    format: date
                    type: string
                id:
                    format: uuid
                    type: string
                notas:
                    nullable: true
                    type: string
                taqueria_uuid:
                    format: uuid
                    type: string
                updated_at:
                    format: date-time
                    type: string
                veredicto:
                    $ref: '#/components/schemas/Veredicto'
            required:
                - id
                - taqueria_uuid
                - fecha
                - veredicto
                - created_at
                - updated_at
            type: object
        VisitaList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/Visita'
                    type: array
                next_page_token:
                    type: string
            type: object