# rest/server/list_visita.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"fmt"
	"log"
	"net/http"

	"go.einride.tech/aip/filtering"
	"go.einride.tech/aip/ordering"

	visitamodule "taqueria/core/module/visita"
	"taqueria/core/module/visita/types"
	main_entity "taqueria/entity/visita"
)

func (h *handlers) ListVisita(w http.ResponseWriter, r *http.Request) {
	defaultPageSize := int32(10)
	maxPageSize := int32(100)

	lp, err := parseListParams(r, defaultPageSize, maxPageSize)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	offset, err := parsePageTokenOffset(lp.pageToken)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid page token: %w", err))
		return
	}

	declarations := visitaDeclarations()
	filter, err := filtering.ParseFilter(lp, declarations)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid filter expression: %w", err))
		return
	}

	orderBy, err := ordering.ParseOrderBy(lp)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid order_by expression: %w", err))
		return
	}

	req := types.ListRequest{
		Offset:                offset,
		PageSize:              lp.GetPageSize(),
		Filter:                filter,
		FilteringDeclarations: declarations,
		EnumValues:            visitaEnumValues(),
		OrderBy:               orderBy,
		IncludeFields:         lp.includeFields,
		ExcludeFields:         lp.excludeFields,
	}

	var result types.ListResponse
	if lp.skipCache {
		result, err = h.core.Visita().List(r.Context(), req, visitamodule.WithSkipCache())
	} else {
		result, err = h.core.Visita().List(r.Context(), req)
	}
	if err != nil {
		// A filter or order_by the clause builder cannot express is the caller's
		// error, not the server's: statusForError turns it into a 400.
		writeProblem(w, statusForError(err), err)
		return
	}

	resp := ListEnvelope[main_entity.Visita]{
		Items: result.Visita,
	}
	if result.HasNextPage {
		resp.NextPageToken = encodePageToken(offset + int64(lp.GetPageSize()))
	}

	writeJSON(w, http.StatusOK, resp)
}

func visitaDeclarations() *filtering.Declarations {
	declarations, err := filtering.NewDeclarations(
		filtering.DeclareStandardFunctions(),
		filtering.DeclareIdent("true", filtering.TypeBool),
		filtering.DeclareIdent("false", filtering.TypeBool),

		// visita

		filtering.DeclareIdent("id", filtering.TypeString),

		filtering.DeclareIdent("taqueria_uuid", filtering.TypeString),

		filtering.DeclareIdent("fecha", filtering.TypeTimestamp),

		// An enum is filtered by its VALUE over REST — either quoted
		// (status = "active", matched case-insensitively against every spelling
		// in visitaEnumValues) or as the generated value name ident
		// (status = STATUS_ACTIVE, the same spelling gRPC accepts). Both are
		// strings to the type checker; the number they stand for is resolved
		// from the value table, not from this declaration.
		filtering.DeclareIdent("veredicto", filtering.TypeString),
		filtering.DeclareIdent("VEREDICTO_INVALID", filtering.TypeString),
		filtering.DeclareIdent("VEREDICTO_IMPERDIBLE", filtering.TypeString),
		filtering.DeclareIdent("VEREDICTO_BUENO", filtering.TypeString),
		filtering.DeclareIdent("VEREDICTO_PASABLE", filtering.TypeString),
		filtering.DeclareIdent("VEREDICTO_NUNCA_MAS", filtering.TypeString),

		filtering.DeclareIdent("notas", filtering.TypeString),

		filtering.DeclareIdent("created_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("updated_at", filtering.TypeTimestamp),
	)
	if err != nil {
		log.Printf("error creating declarations for visita: %v", err)
	}
	return declarations
}

// visitaEnumValues maps each enum filter field to every accepted
// spelling of its values and the number the column stores. REST generates no
// protobuf enum types (an app built with --api rest has no proto package at
// all), so this table — not the filter declarations — is what makes an enum
// filter resolvable here.
func visitaEnumValues() map[string]map[string]int64 {
	return map[string]map[string]int64{

		"veredicto": {
			"invalid":              0,
			"VEREDICTO_INVALID":    0,
			"imperdible":           1,
			"VEREDICTO_IMPERDIBLE": 1,
			"bueno":                2,
			"VEREDICTO_BUENO":      2,
			"pasable":              3,
			"VEREDICTO_PASABLE":    3,
			"nunca_mas":            4,
			"VEREDICTO_NUNCA_MAS":  4,
		},
	}
}